add saving state to the Save changes button
