+++ tutor/src/components/navbar/actions-menu.jsx
@@ -7,7 +7,6 @@ import TourAnchor from '../tours/anchor';
 import Router from '../../helpers/router';
 import UserMenu from '../../models/user/menu';
 import Course from '../../models/course';
-import { breakpoint } from 'theme';
 import Responsive from '../../components/responsive';
 
 const RoutedDropdownItem = (props) => {
